OQAS/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
qrcode==8.2
waitress==3.0.2
Werkzeug==3.1.3
requests==2.31.0
Flask-WTF==1.2.2
